Move seeder validation to new command
